+++ app/src/main/java/com/meten/xyh/service/response/bean/Order.kt
@@ -4,7 +4,7 @@ import com.shuange.lesson.service.response.bean.BaseBean
 
 class Order : BaseBean() {
     var amount: Int = 0
-    var goodsId: Int = 0
+    var goodsId: Long = 0
     var goodsImageUrl: String = ""
     var goodsName: String = ""
     var login: String = ""
@@ -15,5 +15,5 @@ class Order : BaseBean() {
     var refundStatus: String = ""
     var tradeNo: String = ""
     var transactionId: String = ""
-    var userId: Int = 0
+    var userId: Long = 0
 }
