+++ app/src/main/java/com/meten/xyh/service/response/bean/Recharge.kt
@@ -13,5 +13,5 @@ class Recharge : BaseBean() {
     var refundStatus: String = ""
     var tradeNo: String = ""
     var transactionId: String = ""
-    var userId: Int = 0
+    var userId: Long = 0
 }
