+++ app/src/main/java/com/meten/xyh/service/response/bean/UserForAccount.kt
@@ -4,17 +4,17 @@ import com.shuange.lesson.service.response.bean.BaseBean
 
 class UserForAccount : BaseBean() {
     var answer: String = ""
-    var lessonId: Int = 0
+    var lessonId: Long = 0
     var lessonMode: String = ""
-    var lessonModuleId: Int = 0
+    var lessonModuleId: Long = 0
     var lessonModuleName: String = ""
     var lessonName: String = ""
-    var lessonPackageId: Int = 0
+    var lessonPackageId: Long = 0
     var lessonPackageName: String = ""
     var lessonType: String = ""
-    var questionId: Int = 0
+    var questionId: Long = 0
     var score: Int = 0
-    var subUserId: Int = 0
+    var subUserId: Long = 0
     var subUserName: String = ""
-    var userId: Int = 0
+    var userId: Long = 0
 }
