+++ lesson/src/main/java/com/shuange/lesson/service/response/bean/BaseBean.kt
@@ -1,7 +1,7 @@
 package com.shuange.lesson.service.response.bean
 
 open class BaseBean {
-    open val id: Int = 0
+    open val id: Long = 0
     open var createdBy: String = ""
     open var createdDate: String = ""
     open var lastModifiedBy: String = ""
