+++ lesson/src/main/java/com/shuange/lesson/service/response/bean/Module.kt
@@ -2,8 +2,8 @@ package com.shuange.lesson.service.response.bean
 
 data class Module(
     val important: Boolean,
-    val lessonId: Int,
-    val lessonPackageId: Int,
+    val lessonId: Long,
+    val lessonPackageId: Long,
     val lockedStep: Int, //TODO
     val name: String,
     val sortNo: Int
