+++ lesson/src/main/java/com/shuange/lesson/service/response/bean/Record.kt
@@ -1,15 +1,15 @@
 package com.shuange.lesson.service.response.bean
 
 data class Record(
-    val lessonId: Int,
-    val lessonModuleId: Int,
+    val lessonId: Long,
+    val lessonModuleId: Long,
     val lessonModuleName: String,
     val lessonName: String,
-    val lessonPackageId: Int,
+    val lessonPackageId: Long,
     val lessonPackageName: String,
     val lessonType: String,
-    val questionId: Int,
+    val questionId: Long,
     val score: Int,
-    val userId: Int,
+    val userId: Long,
     val userName: String
 ): BaseBean()
